Multi-stage continuous-flow manufacturing process
The data comes from a multi-stage continuous flow manufacturing process. In the first stage, Machines 1, 2, and 3 operate in parallel, and feed their outputs into a step that combines the flows. Output from the combiner is measured in 15 locations surrounding the outer surface of the material exiting the combiner.
PrimaryGoal: Predict measurements of outputfromfirst stage.
Next, the output flows into a second stage, where Machines 4 and 5 process in series. After Machine 5, measurements are made again in the same 15 locations surrounding the outer surface of the material exiting Machine 5.
Secondary Goal: Predict measurements ofoutput fromsecondstage.